added moving to recording page when finished selecting a patient
